backend/pyproject.toml
[tool.poetry]
name = "timekeeper"
version = "0.1.0"
description = ""
authors = ["xpakx <55586470+xpakx@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
sqlalchemy = "^2.0.14"
psycopg2-binary = "^2.9.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
